requirements.txt
# Core ML and Computer Vision
tensorflow>=2.15.0
keras>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Modern Face Detection
mtcnn>=0.1.1
mediapipe>=0.10.0

# Data Processing and Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
plotly>=5.17.0

# Web Interface
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
